+++ Assets/@Project/_Prototypes/RiderAISystem/Chassis/DamageableRiderAI.cs
@@ -14,7 +14,7 @@ namespace DuneRiders.RiderAI {
 			}
 		}
 
-		public override void Damage(int healthPoints)
+		public override void Damage(float healthPoints)
 		{
 			healthState.health -= healthPoints;
 		}
